+++ domino/domino.py
@@ -8,10 +8,12 @@ from domino._version import __version__
 
 import logging
 import requests
+import functools
 from requests.auth import HTTPBasicAuth
 import time
 import pprint
 import re
+import polling2
 from bs4 import BeautifulSoup
 
 
@@ -184,19 +186,8 @@ class Domino:
         return run_response
 
     def run_stop(self, runId, saveChanges=True):
-        """
-        :param runId: string
-        :param saveChanges: boolean (Optional) Save or discard run results.
-        """
-        url = self._routes.job_stop()
-        project_id = self._project_id
-        request = {
-            "projectId": project_id,
-            "jobId": runId,
-            "commitResults": saveChanges
-        }
-        response = self.request_manager.post(url, json=request)
-        return response
+        self.log.warning("Use job_stop method instead")
+        return self.job_stop(job_id=runId, commit_results=saveChanges)
 
     def runs_status(self, runId):
         url = self._routes.runs_status(runId)
@@ -242,6 +233,164 @@ class Domino:
         url = self._routes.runs_stdout(runId)
         # pprint.pformat outputs a string that is ready to be printed
         return pprint.pformat(self._get(url)['stdout'])
+
+    def job_start(self, command: str, commit_id: str = None, hardware_tier_name: str = None,
+                  environment_id: str = None, on_demand_spark_cluster_properties: dict = None) -> dict:
+        """
+        Starts a Domino Job via V4 API
+        :param command:                             string
+                                                    Command to execute in Job
+                                                    >> domino.job_start(command="main.py arg1 arg2")
+        :param commit_id:                           string (Optional)
+                                                    The commitId to launch from. If not provided, will launch
+                                                    from latest commit.
+        :param hardware_tier_name:                  string (Optional)
+                                                    The hardware tier NAME to launch job in. If not provided
+                                                    it will use the default hardware tier for the project
+        :param environment_id:                      string (Optional)
+                                                    The environment id to launch job with. If not provided
+                                                    it will use the default environment for the project
+        :param on_demand_spark_cluster_properties:  dict (Optional)
+                                                    On demand spark cluster properties. Following properties
+                                                    can be provided in spark cluster
+                                                    {
+                                                        "computeEnvironmentId": "<Environment ID configured with spark>"
+                                                        "executorCount": "<Number of Executors in cluster>"
+                                                         (optional defaults to 1)
+                                                        "executorHardwareTierId": "<Hardware tier ID for Spark Executors>"
+                                                         (optional defaults to last used historically if available)
+                                                        "masterHardwareTierId":  "<Hardware tier ID for Spark master"
+                                                         (optional defaults to last used historically if available)
+                                                        "executorStorageMB": "<Executor's storage in MB>"
+                                                         (optional defaults to 0; 1GB is 1000MB Here)
+                                                    }
+        :return: Returns created Job details (number, id etc)
+        """
+        def validate_on_demand_spark_cluster_properties(max_execution_slot_per_user):
+            self.log.debug(f"Validating spark cluster properties: {on_demand_spark_cluster_properties}")
+            validations = {
+                'executorCount': lambda ec: validate_spark_executor_count(int(ec), max_execution_slot_per_user),
+                'computeEnvironmentId': lambda com_env_id: self._validate_environment_id(com_env_id),
+                'executorHardwareTierId': lambda exec_hwd_tier_id: self._validate_hardware_tier_id(exec_hwd_tier_id),
+                'masterHardwareTierId': lambda master_hwd_tier_id: self._validate_hardware_tier_id(master_hwd_tier_id),
+                'executorStorageMB': lambda exec_storage_mb: True
+            }
+            if 'computeEnvironmentId' not in on_demand_spark_cluster_properties:
+                raise Exception(f"Mandatory field computeEnvironmentId not passed in spark properties")
+            for k, v in on_demand_spark_cluster_properties.items():
+                if not validations.get(k, lambda x: False)(v):
+                    raise Exception(f"Invalid spark property {k}:{v} found")
+
+        def validate_spark_executor_count(executor_count, max_executor_count):
+            if executor_count <= max_executor_count:
+                return True
+            else:
+                raise Exception(f"executor count: {executor_count} exceeding max executor count: {max_executor_count}")
+
+        def get_default_spark_settings():
+            self.log.debug(f"Getting default spark settings")
+            default_spark_setting_url = self._routes.default_spark_setting(self._project_id)
+            return self.request_manager.get(default_spark_setting_url).json()
+
+        def validate_is_on_demand_spark_supported():
+            if not is_on_demand_spark_cluster_supported:
+                raise OnDemandSparkClusterNotSupportedException(
+                    f"Your domino deployment version {self._version} does not support on demand spark cluster. "
+                    f"Minimum support version {MINIMUM_ON_DEMAND_SPARK_CLUSTER_SUPPORT_DOMINO_VERSION}")
+
+        spark_cluster_properties = None
+
+        if commit_id is not None:
+            self._validate_commit_id(commit_id)
+        if hardware_tier_name is not None:
+            self._validate_hardware_tier_name(hardware_tier_name)
+        if environment_id is not None:
+            self._validate_environment_id(environment_id)
+        if on_demand_spark_cluster_properties is not None:
+            validate_is_on_demand_spark_supported()
+            default_spark_setting = get_default_spark_settings()
+            max_execution_slot = default_spark_setting['maximumExecutionSlotsPerUser']
+            default_executor_hardware_tier = default_spark_setting['executorHardwareTierId']
+            default_master_hardware_tier = default_spark_setting['masterHardwareTierId']
+            validate_on_demand_spark_cluster_properties(max_execution_slot)
+            executor_hardware_tier_id = on_demand_spark_cluster_properties.get('executorHardwareTierId',
+                                                                               default_executor_hardware_tier)
+            master_hardware_tier_id = on_demand_spark_cluster_properties.get('masterHardwareTierId',
+                                                                             default_master_hardware_tier)
+            executor_count = int(on_demand_spark_cluster_properties.get('executorCount', 1))
+            compute_environment_id = on_demand_spark_cluster_properties.get('computeEnvironmentId')
+            executor_storage = int(on_demand_spark_cluster_properties.get('executorStorageMB', 0))
+            spark_cluster_properties = {
+                "computeEnvironmentId": compute_environment_id,
+                "executorCount": executor_count,
+                "executorHardwareTierId": executor_hardware_tier_id,
+                "executorStorageMB": executor_storage,
+                "masterHardwareTierId": master_hardware_tier_id
+            }
+
+        url = self._routes.job_start()
+        payload = {
+          "projectId": self._project_id,
+          "commandToRun": command,
+          "commitId": commit_id,
+          "overrideHardwareTierName": hardware_tier_name,
+          "onDemandSparkClusterProperties": spark_cluster_properties,
+          "environmentId": environment_id
+        }
+        response = self.request_manager.post(url, json=payload)
+        return response.json()
+
+    def job_stop(self, job_id: str, commit_results: bool = True):
+        """
+        Stops the Job with given job_id
+        :param job_id: The job identifier
+        :param commit_results: Boolean representing to commit results to file
+        """
+        url = self._routes.job_stop()
+        request = {
+            "projectId": self._project_id,
+            "jobId": job_id,
+            "commitResults": commit_results
+        }
+        response = self.request_manager.post(url, json=request)
+        return response
+
+    def job_status(self, job_id: str) -> dict:
+        """
+        Gets the status of job with given job_id
+        :param job_id: The job identifier
+        :return: The details
+        """
+        return self.request_manager.get(
+            self._routes.job_status(job_id)
+        ).json()
+
+    def job_start_blocking(self, poll_freq: int = 5, max_poll_time: int = 6000, **kwargs) -> dict:
+        """
+        Starts a job in a blocking loop, periodically polling for status of job
+        is complete. Will ignore intermediate request exception.
+        :param poll_freq: int (Optional) polling frequency interval in seconds
+        :param max_poll_time: int (Optional) max poll time in seconds
+        :param kwargs: Additional arguments to be passed to job_start
+        :return:
+        """
+        def get_job_status(job_identifier):
+            status = self.job_status(job_identifier)
+            self.log.info(f"Polling Job: {job_identifier} status is completed: {status['statuses']['isCompleted']}")
+            return status
+        job = self.job_start(**kwargs)
+        job_id = job['id']
+        job_status = polling2.poll(
+            target=lambda: get_job_status(job_id),
+            check_success=lambda status: status['statuses']['isCompleted'],
+            ignore_exceptions=(requests.exceptions.RequestException,),
+            timeout=max_poll_time,
+            step=poll_freq,
+            log_error=self.log.level
+        )
+        stdout_msg = self.get_run_log(runId=job_id, includeSetupLog=False)
+        self.process_log(stdout_msg)
+        return job_status
 
     def files_list(self, commitId, path='/'):
         url = self._routes.files_list(commitId, path)
@@ -455,6 +604,40 @@ class Domino:
                 else:
                     self.log.info(text)
 
+    # Validation methods
+    def _useable_environments_list(self):
+        self.log.debug(f"Getting list of useable environment")
+        useable_environment_list_url = self._routes.useable_environments_list(self._project_id)
+        return self.request_manager.get(useable_environment_list_url).json()['environments']
+
+    def _validate_environment_id(self, environment_id):
+        self.log.debug(f"Validating environment id: {environment_id}")
+        for environment in self._useable_environments_list():
+            if environment_id == environment['id']:
+                return True
+        raise EnvironmentNotFoundException(f"{environment_id} environment not found")
+
+    def _validate_hardware_tier_id(self, hardware_tier_id):
+        self.log.debug(f"Validating hardware tier id: {hardware_tier_id}")
+        for hardware_tier in self.hardware_tiers_list():
+            if hardware_tier_id == hardware_tier['hardwareTier']['id']:
+                return True
+        raise HardwareTierNotFoundException(f"{hardware_tier_id} hardware tier Id not found")
+
+    def _validate_hardware_tier_name(self, hardware_tier_name):
+        self.log.debug(f"Validating hardware tier id: {hardware_tier_name}")
+        for hardware_tier in self.hardware_tiers_list():
+            if hardware_tier_name == hardware_tier['hardwareTier']['name']:
+                return True
+        raise HardwareTierNotFoundException(f"{hardware_tier_name} hardware tier name not found")
+
+    def _validate_commit_id(self, commit_id):
+        self.log.debug(f"Validating commit id: {commit_id}")
+        for commit in self.commits_list():
+            if commit_id == commit:
+                return True
+        raise CommitNotFoundException(f"{commit_id} commit Id not found")
+
     # Helper methods
     def _get(self, url):
         return self.request_manager.get(url).json()
@@ -479,6 +662,7 @@ class Domino:
 
     # Workaround to get project ID which is needed for some model functions
     @property
+    @functools.lru_cache()
     def _project_id(self):
         url = self._routes.find_project_by_owner_name_and_project_name_url()
         key = "id"